+++ example/input_4.dart
@@ -1,6 +1,6 @@
 import 'package:rpi_gpio/rpi_gpio.dart';
 
-import 'example_app.dart';
+import 'input_4_app.dart';
 
 /// Launch the example by injecting an instance of [RpiGpio].
 /// This only works on the Raspberry Pi.
@@ -10,5 +10,5 @@ import 'example_app.dart';
 /// and thus can be tested with a mock [Gpio] on any platform.
 void main() async {
   final gpio = await initialize_RpiGpio();
-  await runExample(gpio);
+  await runInput4(gpio);
 }