+++ test/all.dart
@@ -6,7 +6,7 @@ import 'example_test.dart' as example;
 import 'src/polling_node_test.dart' as polling_node;
 import 'src/pwm_node_test.dart' as pwm_node;
 
-main() {
+void main() {
   group('polling', () => polling_node.main());
   group('pwm', () => pwm_node.main());
   group('basic', () => basic.main());